+++ shared_libraries/bin/run_tests.sh
@@ -5,23 +5,24 @@
 ##
 ## DESCRIPTION
 ## -----------------------------------------------------------------------------
-## Tests the lambda (task) for post_copy_request_to_queue using unit tests.
+## Tests the shared library for shared_db using unit tests.
 ##
 ##
 ## USAGE
 ## -----------------------------------------------------------------------------
 ## bin/run_tests.sh
 ##
-## This must be called from the (root) lambda directory /tasks/post_copy_request_to_queue
+## This must be called from the (root) shared library directory
+## tasks/shared_libraries
 ## =============================================================================
 
 ## Set this for Debugging only
-#set -ex
+#set -x
 
 ## Make sure we are calling the script the correct way.
 BASEDIR=$(dirname $0)
 if [ "$BASEDIR" != "bin" ]; then
-  >&2 echo "ERROR: This script must be called from the root directory of the task lambda [bin/run_tests.sh]."
+  >&2 echo "ERROR: This script must be called from the root directory of shared_libraries [bin/run_tests.sh]."
   exit 1
 fi
 
@@ -61,22 +62,22 @@ source venv/bin/activate
 
 ## Install the requirements
 pip install -q --upgrade pip --trusted-host pypi.org --trusted-host files.pythonhosted.org
-pip install -q -r requirements-dev.txt --trusted-host pypi.org --trusted-host files.pythonhosted.org
+pip install -q -r requirements.txt --trusted-host pypi.org --trusted-host files.pythonhosted.org
 let return_code=$?
 
 check_rc $return_code "ERROR: pip install encountered an error."
 
-## copy the shared_recovery.py
-echo "INFO: Copying ORCA shared libraries ..."
-if [ -d orca_shared ]; then
-    rm -rf orca_shared
-fi
-
 ## Run unit tests and check Coverage
 echo "INFO: Running unit and coverage tests ..."
 
-# Currently just running unit tests until we fix/support large tests
-coverage run --source post_copy_request_to_queue -m pytest
+# Export the AWS_REGION for the boto3 clients. Note that if you clear the
+# environment for unit tests, you will need to add this environment variable back
+# in to avoid a client error "botocore.exceptions.NoRegionError: You must specify
+# a region." In real AWS, the AWS_REGION is set for you per the Lamda developer
+# docs seen here https://docs.aws.amazon.com/lambda/latest/dg/lambda-dg.pdf
+export AWS_REGION="us-west-2"
+
+coverage run --source=orca_shared -m pytest
 let return_code=$?
 check_rc $return_code "ERROR: Unit tests encountered failures."
 
@@ -85,11 +86,48 @@ coverage report --fail-under=80
 let return_code=$?
 check_rc $return_code "ERROR: Unit tests coverage is less than 80%"
 
+
+## Run code smell and security tests using bandit
+echo "INFO: Running code smell security tests ..."
+bandit -r orca_shared
+let return_code=$?
+check_rc $return_code "ERROR: Potential security or code issues found."
+
+
+## Check code third party libraries for CVE issues
+echo "INFO: Running checks on third party libraries ..."
+safety check -r requirements.txt
+let return_code=$?
+check_rc $return_code "ERROR: Potential security issues third party libraries."
+
+
+## Check code formatting and styling
+echo "INFO: Checking formatting and style of code ..."
+echo "INFO: Checking lint rules ..."
+flake8 \
+    --max-line-length 99 \
+    --extend-ignore E203 \
+    orca_shared
+check_rc $return_code "ERROR: Linting issues found."
+
+echo "INFO: Sorting imports ..."
+isort \
+    --trailing-comma \
+    --ensure-newline-before-comments \
+    --line-length 88 \
+    --use-parentheses \
+    --force-grid-wrap 0 \
+    -m 3 \
+    orca_shared
+
+echo "INFO: Formatting with black ..."
+black orca_shared
+
+
 ## Deactivate and remove the virtual env
 echo "INFO: Cleaning up the environment ..."
 deactivate
 rm -rf venv
 find . -type d -name "__pycache__" -exec rm -rf {} +
-# Remove the include dir from greenlet sqlalchemy
-rm -rf include
+
 exit 0