+++ tasks/post_copy_request_to_queue/bin/build.sh
@@ -82,28 +82,6 @@ check_rc $return_code "ERROR: pip install encountered an error."
 # Install the aws-lambda psycopg2 libraries
 mkdir -p build/psycopg2
 
-## copy the shared_recovery.py
-echo "INFO: Copying ORCA shared libraries ..."
-if [ -d orca_shared ]; then
-    rm -rf orca_shared
-fi
-
-mkdir -p build/orca_shared
-let return_code=$?
-check_rc $return_code "ERROR: Unable to create orca_shared directory."
-
-touch build/orca_shared/__init__.py
-let return_code=$?
-check_rc $return_code "ERROR: Unable to create [orca_shared/__init__.py] file"
-
-cp ../shared_libraries/recovery/shared_recovery.py build/orca_shared/
-let return_code=$?
-check_rc $return_code "ERROR: Unable to copy shared library [orca_shared/shared_recovery.py]"
-
-cp ../shared_libraries/database/shared_db.py build/orca_shared/
-let return_code=$?
-check_rc $return_code "ERROR: Unable to copy shared library [orca_shared/shared_db.py]"
-
 ##TODO: Adjust build scripts to put shared packages needed under a task/build/packages directory.
 ##      and copy the packages from there.
 if [ ! -d "../package" ]; then
@@ -146,7 +124,5 @@ rm -rf build
 rm -rf include
 rm -rf venv
 find . -type d -name "__pycache__" -exec rm -rf {} +
-# Remove the shared library
-rm -rf orca_shared
 
 exit 0